Crate engine_panic

An example TiKV storage engine.

This project is intended to serve as a skeleton for other engine implementations. It lays out the complex system of engine modules and traits in a way that is consistent with other engines. To create a new engine simply copy the entire directory structure and replace all "Panic*" names with your engine's own name; then fill in the implementations; remove the allow(unused) attribute;

Re-exports

pub use crate::range_properties::*;

Modules

cf_handle

cf_names

cf_options

compact

db_options

db_vector

engine

import

misc

range_properties

snapshot

sst

table_properties

write_batch

Structs

PanicCFHandle

PanicColumnFamilyOptions

PanicCompactedEvent

PanicDBOptions

PanicDBVector

PanicEngine

PanicEngineIterator

PanicExternalSstFileInfo

PanicExternalSstFileReader

PanicIngestExternalFileOptions

PanicSnapshot

PanicSnapshotIterator

PanicSstReader

PanicSstReaderIterator

PanicSstWriter

PanicSstWriterBuilder

PanicTableProperties

PanicTablePropertiesCollection

PanicTablePropertiesCollectionIter

PanicTablePropertiesKey

PanicTitanDBOptions

PanicUserCollectedProperties

PanicWriteBatch